+++ src/rpc/blockchain.cpp
@@ -13,6 +13,7 @@
 #include "consensus/validation.h"
 #include "validation.h"
 #include "core_io.h"
+#include "policy/feerate.h"
 #include "policy/policy.h"
 #include "primitives/transaction.h"
 #include "rpc/server.h"
