+++ src/test/amount_tests.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "amount.h"
+#include "policy/feerate.h"
 #include "test/test_bitcoin.h"
 
 #include <boost/test/unit_test.hpp>
