+++ src/wallet/coincontrol.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_WALLET_COINCONTROL_H
 #define BITCOIN_WALLET_COINCONTROL_H
 
+#include "policy/feerate.h"
 #include "primitives/transaction.h"
 #include "wallet/wallet.h"
